az-function/requirements.txt
azure-functions==1.17.0
httpx==0.27.2
httpcore[asyncio]>=1.0.4,<2.0.0
anyio>=4.4.0,<5.0.0
sniffio>=1.3.0,<2.0.0
pydantic==2.9.2
